1.0

+++ SIGMA_PRUEBA/Models/AlumnosParams.cs
@@ -5,8 +5,8 @@ namespace SIGMA_PRUEBA.Models
 public class AlumnosParams
 {
     // Default Id.
-    public int Id { get; set; }
-    public int Codigo { get; set; }
+    public long Id { get; set; }
+    public long Codigo { get; set; }
     public string Nombre { get; set; }
     public string ApellidoP { get; set; }
     public string ApellidoM { get; set; }
